Added wandb to requirement, moved ToxicCommentClassifier class into model.py and removed it from train_model.py (but inserted as an import)
