+++ requirements.txt
@@ -7,3 +7,4 @@ scikit_learn==1.3.2
 torch==2.1.2
 transformers==4.36.2
 tensorboard==2.15.1
+wandb==0.16.2
